Adjust backport to work with stable9.1

+++ apps/files_trashbin/lib/BackgroundJob/ExpireTrash.php
@@ -76,14 +76,25 @@ class ExpireTrash extends \OC\BackgroundJob\TimedJob {
 			return;
 		}
 
-		$this->userManager->callForSeenUsers(function(IUser $user) {
-			$uid = $user->getUID();
-			if (!$this->setupFS($uid)) {
-				return;
-			}
-			$dirContent = Helper::getTrashFiles('/', $uid, 'mtime');
-			Trashbin::deleteExpiredFiles($dirContent, $uid);
-		});
+		if (is_callable(array($this->userManager, 'callForSeenUsers'))) {
+			$this->userManager->callForSeenUsers(function(IUser $user) {
+				$uid = $user->getUID();
+				if (!$this->setupFS($uid)) {
+					return;
+				}
+				$dirContent = Helper::getTrashFiles('/', $uid, 'mtime');
+				Trashbin::deleteExpiredFiles($dirContent, $uid);
+			});
+		} else {
+			$this->userManager->callForAllUsers(function(IUser $user) {
+				$uid = $user->getUID();
+				if ($user->getLastLogin() === 0 || !$this->setupFS($uid)) {
+					return;
+				}
+				$dirContent = Helper::getTrashFiles('/', $uid, 'mtime');
+				Trashbin::deleteExpiredFiles($dirContent, $uid);
+			});
+		}
 		
 		\OC_Util::tearDownFS();
 	}
